catweb/requirements.txt
Cerberus==1.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
ldap3==2.5.2
MarkupSafe==1.1.1
numpy==1.16.2
pandas==0.24.1
passlib==1.7.1
psutil==5.5.1
pyasn1==0.4.5
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
